# etc/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Homer-App User interface for WEB AI",
    "title": "Homer-App",
    "license": {
      "name": "AGPL",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "1.1.2"
  },
  "host": "localhost:9080",
  "basePath": "/api/v3",
  "paths": {
...
    "/call/report/log": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns log data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchGetTransactionLog",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SearchTransactionLogList",
            "schema": {
              "$ref": "#/definitions/SearchTransactionLogList"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/call/report/qos": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns qos data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchGetTransactionQos",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SearchTransactionQos",
            "schema": {
              "$ref": "#/definitions/SearchTransactionQos"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/call/transaction": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns transaction data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchGetTransaction",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SearchTransaction",
            "schema": {
              "$ref": "#/definitions/SearchTransaction"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
...
    "/export/call/messages/pcap": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns pcap data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/octet-stream"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchGetMessagesAsPCap",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PCAPResponse",
            "schema": {
              "$ref": "#/definitions/PCAPResponse"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/export/call/messages/text": {
      "post": {
        "security": [
          {
            "bearer": []
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns text data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/octet-stream"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchGetMessagesAsText",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TextResponse",
            "schema": {
              "$ref": "#/definitions/TextResponse"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
...
    "/import/data/pcap": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns pcap data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Import"
        ],
        "operationId": "GetMessagesAsPCap",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ListUsers",
            "schema": {
              "$ref": "#/definitions/ListUsers"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/import/data/pcap/now": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns pcap data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Import"
        ],
        "operationId": "GetMessagesAsPCapNow",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ListUsers",
            "schema": {
              "$ref": "#/definitions/ListUsers"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
...
    "/search/call/data": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchSearchData",
        "parameters": [
          {
            "type": "object",
            "description": "SearchTransactionRequest parameters",
            "name": "SearchTransactionRequest",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchTransactionRequest parameters",
              "type": "object",
              "$ref": "#/definitions/SearchTransactionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SearchCallData",
            "schema": {
              "$ref": "#/definitions/SearchCallData"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/search/call/decode/message": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchGetDecodeMessageById",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MessageDecoded",
            "schema": {
              "$ref": "#/definitions/MessageDecoded"
            }
          },
          "400": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/search/call/message": {
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Returns message data based upon filtered json",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "search"
        ],
        "operationId": "searchGetMessageById",
        "parameters": [
          {
            "type": "object",
            "description": "SearchObject parameters",
            "name": "SearchObject",
            "in": "body",
            "required": true,
            "schema": {
              "description": "SearchObject parameters",
              "type": "object",
              "$ref": "#/definitions/SearchObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SearchCallData",
            "schema": {
              "$ref": "#/definitions/SearchCallData"
            }
          },
          "400": {
            "description": "UserLoginFailureResponse",
            "schema": {
              "$ref": "#/definitions/UserLoginFailureResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SearchObject": {
      "type": "object",
      "required": [
        "param"
      ],
      "properties": {
        "param": {
          "type": "object",
          "required": [
            "limit"
          ],
          "properties": {
            "limit": {
              "description": "this controls the number of records to display",
              "type": "integer",
              "format": "int64",
              "x-go-name": "Limit",
              "example": 200
            },
            "location": {
              "description": "location",
              "type": "object",
              "properties": {
                "node": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "x-go-name": "Node"
                }
              },
              "x-go-name": "Location"
            },
            "orlogic": {
              "description": "this control the type of search one can perform\ntype: boolean",
              "type": "boolean",
              "x-go-name": "OrLogic",
              "example": false
            },
            "search": {
              "description": "this control the type of search one can perform\ntype: string\nformat: binary",
              "type": "object",
              "x-go-name": "Search",
              "example": "`{\"1_call\":[{\"name\":\"limit\",\"value\":\"10\",\"type\":\"string\",\"hepid\":1}]}`"
            },
            "timezone": {
              "description": "timezone settings\ntype: object",
              "type": "object",
              "default": null,
              "properties": {
                "name": {
                  "type": "string",
                  "x-go-name": "Name"
                },
                "value": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Value"
                }
              },
              "x-go-name": "Timezone"
            },
            "transaction": {
              "type": "object",
              "x-go-name": "Transaction"
            },
            "whitelist": {
              "description": "ips to be removed from search",
              "type": "array",
              "items": {
                "type": "string"
              },
              "x-go-name": "WhiteList",
              "example": [
                "192.1698.10.20"
              ]
            }
          },
          "x-go-name": "Param"
        },
        "timestamp": {
          "description": "this control the time range for used for search",
          "type": "object",
          "required": [
            "from",
            "to"
          ],
          "properties": {
            "from": {
              "description": "current timestamp in miliseconds",
              "type": "integer",
              "format": "int64",
              "x-go-name": "From",
              "example": 1581793200000
            },
            "to": {
              "description": "current timestamp in miliseconds",
              "type": "integer",
              "format": "int64",
              "x-go-name": "To",
              "example": 1581879599000
            }
          },
          "x-go-name": "Timestamp"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
    "SearchTransactionLogList": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SearchTransactionLog"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}